download assets

+++ addons/minecraft_launcher/utils.gd
@@ -18,7 +18,7 @@ static func check_sha1(file_path: String, sha1: String, context: int = HashingCo
 	var res = ctx.finish()
 	return res.hex_encode() == sha1
 
-static func download_file(http_request: HTTPRequest, url: String, path: String, sha1: String) -> void:
+static func download_file(http_request: HTTPRequest, url: String, path: String, sha1: String = "") -> void:
 	DirAccess.make_dir_recursive_absolute(path.replace(path.get_file(), "")) # TODO: check err
 	# if something is missing, don't do it
 	if url == "" || path == "": return
@@ -29,7 +29,7 @@ static func download_file(http_request: HTTPRequest, url: String, path: String, 
 		await http_request.request_completed
 	
 	# check if file is correct
-	if not Utils.check_sha1(path, sha1):
+	if sha1 != "" and not Utils.check_sha1(path, sha1):
 		DirAccess.remove_absolute(path)
 
 static func download_content_file(http_request: HTTPRequest) -> Variant:
